requirements.txt
flask==3.1.2
flask-mail==0.10.0
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
flask-login==0.6.3
flask-wtf==1.2.2
email-validator==2.3.0
bcrypt==5.0.0
werkzeug==3.1.5
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-api-python-client==2.188.0
python-docx==1.1.2
Pillow==12.1.1

# Production
gunicorn==23.0.0
psycopg2-binary==2.9.10
boto3==1.38.0
